Introduce Makefile and correct go.mod
This commit introduces a Makefile which can be used to drive tasks from
testing to running cragcast. Running `make` by itself will print help
text to the developer revealing commands.

Additionally this commit fixes the go.mod structure of the project.
There should only be 1 go.mod for the entire project and the packages
within the project should be able to import other packages without the
complexity of managing their version in the go.mod. This includes
remove the need to do "require"/"replace" style imports. This now
just works when inside the same repository.