Introduce weather-client interface and package
This commit introduces the weather-client interface and package. The
interface enables abstraction of Noaa from our other packages (and APIs)
such that we can expand over time how weather data is retrieved. It also
enables mocking for importers of weather-client to write unit tests that
don't need to reach out to external services.

Along with the interface introduction, the noaa-client package is now
renamed weather-client. This change is to support the generic nature in
which weather is retrieved. For now, noaa's implementation of the
WeatherClient interface should remain in the weather-client package.
However, over time, a case could be made that the noaa-client should be
refactored into its own.